+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # filepaths
 
+## Unreleased
+
+#### Fixed
+
+- Support `*` (`:wild`) in extension position.
+
 ## 1.0.0 (2025-02-15)
 
 Finalising `1.0.0` since it's been stable for over a year without issues.